apps/weather_cli/requirements.txt
inquirer
rich
requests
python-dotenv